Rename components and remove complexity from gulp task
