+++ .gitignore
@@ -1,5 +1,6 @@
 dist
 node_modules
 npm-debug.log
+npm-constructicon
 npm-ps-components
 styleguide
