+++ gulpfile.js
@@ -19,22 +19,12 @@ var otherFiles = [
   '!./dist/*.js'
 ]
 
-var camelCaseToDashCase = function (string) {
-  return string.replace(/([a-z])([A-Z])/g, '$1-$2').toLowerCase()
-}
-
 /**
 * Place components straight into the root directory and update references
 */
 gulp.task('npm-prep-components', function() {
   return gulp.src(componentsFiles)
     .pipe(replace('../..', '..'))
-    .pipe(rename(function (path) {
-      path.dirname = camelCaseToDashCase(path.dirname)
-    }))
-    .pipe(replace(/require\(\'\.\.(.*)/g, function(match) {
-      return camelCaseToDashCase(match)
-    }))
     .pipe(gulp.dest(npmDir))
 })
 
@@ -44,8 +34,7 @@ gulp.task('npm-prep-components', function() {
 gulp.task('npm-prep-root', function() {
   return gulp.src(rootFiles)
     .pipe(replace(/.\/components\/(.*)/g, function(match) {
-      var name = match.replace(/.\/components\//g, './');
-      return camelCaseToDashCase(name)
+      return match.replace(/.\/components\//g, './');
     }))
     .pipe(gulp.dest(npmDir))
 })
