+++ source/components/Leaderboard/index.js
@@ -1,5 +1,5 @@
 import React, { Component, PropTypes } from 'react'
-import Icon from '../Icon'
+import Icon from '../icon'
 import { withStyles } from '../../lib/css'
 import options from '../../lib/traits/options'
 import styles from './styles'
