+++ source/components/Modal/index.js
@@ -1,6 +1,6 @@
 import React, { PropTypes } from 'react'
 import ReactModal from 'react-modal'
-import Icon from '../Icon'
+import Icon from '../icon'
 import { withStyles } from '../../lib/css'
 import styles from './styles'
