+++ source/components/button-group/__tests__/ButtonGroup-test.js
@@ -1,5 +1,5 @@
 import ButtonGroup from '..'
-import Button from '../../Button'
+import Button from '../../button'
 import { css } from '../../../lib/css'
 
 describe('Button Group', () => {
