+++ source/components/button-social/index.js
@@ -1,6 +1,6 @@
 import React, { PropTypes } from 'react'
-import Button from '../Button'
-import Icon from '../Icon'
+import Button from '../button'
+import Icon from '../icon'
 import openShareDialog from './openShareDialog'
 
 /**
