+++ source/components/leaderboard-item/styles.js
