+++ source/components/rich-text/__tests__/RichText-test.js
