+++ source/components/rich-text/index.js
