+++ source/components/rich-text/styles.js
