+++ source/components/traits-provider/index.js
